+++ base/linalg/linalg.jl
@@ -22,6 +22,9 @@ export
 
 # Types
     RowVector,
+    ConjArray,
+    ConjVector,
+    ConjMatrix,
     SymTridiagonal,
     Tridiagonal,
     Bidiagonal,
@@ -237,6 +240,7 @@ end
 copy_oftype{T,N}(A::AbstractArray{T,N}, ::Type{T}) = copy(A)
 copy_oftype{T,N,S}(A::AbstractArray{T,N}, ::Type{S}) = convert(AbstractArray{S,N}, A)
 
+include("conjarray.jl")
 include("transpose.jl")
 include("rowvector.jl")