+++ base/linalg/rowvector.jl
@@ -1,17 +1,15 @@
 """
     RowVector(vector)
 
-A lazy-view wrapper of an `AbstractVector`, which turns a length-`n` vector into
-a `1×n` shaped row vector and represents the transpose of a vector (the elements
-are also transposed recursively). This type is usually constructed (and
-unwrapped) via the `transpose()` function or `.'` operator (or related
-`ctranspose()` or `'` operator).
+A lazy-view wrapper of an `AbstractVector`, which turns a length-`n` vector into a `1×n`
+shaped row vector and represents the transpose of a vector (the elements are also transposed
+recursively). This type is usually constructed (and unwrapped) via the [`transpose`](@ref)
+function or `.'` operator (or related [`ctranspose`](@ref) or `'` operator).
 
-By convention, a vector can be multiplied by a matrix on its left (`A * v`)
-whereas a row vector can be multiplied by a matrix on its right (such that
-`v.' * A = (A.' * v).'`). It differs from a `1×n`-sized matrix by the facts that
-its transpose returns a vector and the inner product `v1.' * v2` returns a
-scalar, but will otherwise behave similarly.
+By convention, a vector can be multiplied by a matrix on its left (`A * v`) whereas a row
+vector can be multiplied by a matrix on its right (such that `v.' * A = (A.' * v).'`). It
+differs from a `1×n`-sized matrix by the facts that its transpose returns a vector and the
+inner product `v1.' * v2` returns a scalar, but will otherwise behave similarly.
 """
 immutable RowVector{T,V<:AbstractVector} <: AbstractMatrix{T}
     vec::V
@@ -21,10 +19,11 @@ immutable RowVector{T,V<:AbstractVector} <: AbstractMatrix{T}
     end
 end
 
-
 @inline check_types{T1,T2}(::Type{T1},::AbstractVector{T2}) = check_types(T1, T2)
 @pure check_types{T1,T2}(::Type{T1},::Type{T2}) = T1 === transpose_type(T2) ? nothing :
     error("Element type mismatch. Tried to create a `RowVector{$T1}` from an `AbstractVector{$T2}`")
+
+typealias ConjRowVector{T, CV <: ConjVector} RowVector{T, CV}
 
 # The element type may be transformed as transpose is recursive
 @inline transpose_type{T}(::Type{T}) = promote_op(transpose, T)
@@ -47,10 +46,12 @@ end
 convert{T,V<:AbstractVector}(::Type{RowVector{T,V}}, rowvec::RowVector) =
     RowVector{T,V}(convert(V,rowvec.vec))
 
-# similar()
-@inline similar(rowvec::RowVector) = RowVector(similar(rowvec.vec))
-@inline similar{T}(rowvec::RowVector, ::Type{T}) = RowVector(similar(rowvec.vec, transpose_type(T)))
-# There is no resizing similar() because it would be ambiguous if the result were a Matrix or a RowVector
+# similar tries to maintain the RowVector wrapper and the parent type
+@inline similar(rowvec::RowVector) = RowVector(similar(parent(rowvec)))
+@inline similar{T}(rowvec::RowVector, ::Type{T}) = RowVector(similar(parent(rowvec), transpose_type(T)))
+
+# Resizing similar currently loses its RowVector property.
+@inline similar{T,N}(rowvec::RowVector, ::Type{T}, dims::Dims{N}) = similar(parent(rowvec), T, dims)
 
 # Basic methods
 """
@@ -73,18 +74,34 @@ julia> transpose(v)
 ```
 """
 @inline transpose(vec::AbstractVector) = RowVector(vec)
-@inline ctranspose{T}(vec::AbstractVector{T}) = RowVector(conj(vec))
+@inline ctranspose{T}(vec::AbstractVector{T}) = RowVector(_conj(vec))
 @inline ctranspose{T<:Real}(vec::AbstractVector{T}) = RowVector(vec)
 
 @inline transpose(rowvec::RowVector) = rowvec.vec
+@inline transpose(rowvec::ConjRowVector) = copy(rowvec.vec) # remove the ConjArray wrapper from any raw vector
 @inline ctranspose{T}(rowvec::RowVector{T}) = conj(rowvec.vec)
 @inline ctranspose{T<:Real}(rowvec::RowVector{T}) = rowvec.vec
 
 parent(rowvec::RowVector) = rowvec.vec
 
-# Strictly, these are unnecessary but will make things stabler if we introduce
-# a "view" for conj(::AbstractArray)
-@inline conj(rowvec::RowVector) = RowVector(conj(rowvec.vec))
+"""
+    conj(rowvector)
+
+Returns a [`ConjArray`](@ref) lazy view of the input, where each element is conjugated.
+
+### Example
+
+```jldoctest
+julia> v = [1+im, 1-im].'
+1×2 RowVector{Complex{Int64},Array{Complex{Int64},1}}:
+ 1+1im  1-1im
+
+julia> conj(v)
+1×2 RowVector{Complex{Int64},ConjArray{Complex{Int64},1,Array{Complex{Int64},1}}}:
+ 1-1im  1+1im
+```
+"""
+@inline conj(rowvec::RowVector) = RowVector(_conj(rowvec.vec))
 @inline conj{T<:Real}(rowvec::RowVector{T}) = rowvec
 
 # AbstractArray interface
@@ -147,6 +164,11 @@ end
 
 # Multiplication #
 
+# inner product -> dot product specializations
+@inline *{T<:Real}(rowvec::RowVector{T}, vec::AbstractVector{T}) = dot(parent(rowvec), vec)
+@inline *(rowvec::ConjRowVector, vec::AbstractVector) = dot(rowvec', vec)
+
+# Generic behavior
 @inline function *(rowvec::RowVector, vec::AbstractVector)
     if length(rowvec) != length(vec)
         throw(DimensionMismatch("A has dimensions $(size(rowvec)) but B has dimensions $(size(vec))"))